[Auto Generated] 3.0.0-beta.6

+++ Sources/code/platform/Models/Order/ResponseDetailOrderPlatformModel.swift
@@ -16,7 +16,7 @@ public extension PlatformClient.Order {
         
         public var message: String?
         
-        public var status: Int?
+        public var status: Double?
         
 
         public enum CodingKeys: String, CodingKey {
@@ -29,7 +29,7 @@ public extension PlatformClient.Order {
             
         }
 
-        public init(message: String? = nil, status: Int? = nil, success: Bool? = nil) {
+        public init(message: String? = nil, status: Double? = nil, success: Bool? = nil) {
             
             self.success = success
             
@@ -68,7 +68,7 @@ public extension PlatformClient.Order {
             
             
                 do {
-                    status = try container.decode(Int.self, forKey: .status)
+                    status = try container.decode(Double.self, forKey: .status)
                 
                 } catch DecodingError.typeMismatch(let type, let context) {
                     print("Type '\(type)' mismatch:", context.debugDescription)
@@ -118,7 +118,7 @@ public extension PlatformClient.ApplicationClient.Order {
         
         public var message: String?
         
-        public var status: Int?
+        public var status: Double?
         
 
         public enum CodingKeys: String, CodingKey {
@@ -131,7 +131,7 @@ public extension PlatformClient.ApplicationClient.Order {
             
         }
 
-        public init(message: String? = nil, status: Int? = nil, success: Bool? = nil) {
+        public init(message: String? = nil, status: Double? = nil, success: Bool? = nil) {
             
             self.success = success
             
@@ -170,7 +170,7 @@ public extension PlatformClient.ApplicationClient.Order {
             
             
                 do {
-                    status = try container.decode(Int.self, forKey: .status)
+                    status = try container.decode(Double.self, forKey: .status)
                 
                 } catch DecodingError.typeMismatch(let type, let context) {
                     print("Type '\(type)' mismatch:", context.debugDescription)
